requirements.txt
pandas
numpy
matplotlib
seaborn
sklearn
biopython
